requirements.txt
panel>=1.2.0
param>=2.0.0
langchain>=0.0.300
langchain-community>=0.0.10
langchain-openai>=0.0.3
openai>=1.3.0
chromadb>=0.4.14
flask>=2.0.0
python-dotenv>=0.19.0
langchain-chroma>=0.0.5
